Changed structure of the project
To centre the markup package around the repo root.
