+++ cmd/cmds/completion.go
@@ -1,4 +1,4 @@
-package cmd
+package cmds
 
 import (
 	"fmt"
