+++ cmd/cmds/root.go
@@ -1,0 +1,31 @@
+package cmds
+
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+	"os"
+)
+
+var cfgFile string
+
+// rootCmd represents the base command when called without any subcommands
+var rootCmd = &cobra.Command{
+	Use:   "markup",
+	Short: "A tool to facilitate the rendering of PDFs using Pandoc and LaTeX templates",
+	/*Long: `A longer description that spans multiple lines and likely contains
+examples and usage of using your application. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra application.`,*/
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
+
