+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/IdlePhysicist/markup/cmd/cmds"
+
+func main() {
+	cmds.Execute()
+}
